Refactor the animation of the parameters card
